[Export] Put template engine into own file
 - new SimpleEngine which is the current one
 - MediaExport is responsible for selecting the correct engine
