+++ MediaElch.pro
@@ -141,6 +141,7 @@ SOURCES += src/main.cpp \
     src/export/ExportTemplate.cpp \
     src/export/ExportTemplateLoader.cpp \
     src/export/MediaExport.cpp \
+    src/export/SimpleEngine.cpp \
     src/file/FileFilter.cpp \
     src/globals/Actor.cpp \
     src/globals/ComboDelegate.cpp \
@@ -384,6 +385,7 @@ HEADERS  += Version.h \
     src/export/ExportTemplate.h \
     src/export/ExportTemplateLoader.h \
     src/export/MediaExport.h \
+    src/export/SimpleEngine.h \
     src/file/FileFilter.h \
     src/globals/Actor.h \
     src/globals/ComboDelegate.h \
