+++ test/integration/export/testSimpleExport.cpp
@@ -1,0 +1,8 @@
+#include "test/test_helpers.h"
+
+#include "export/MediaExport.h"
+
+TEST_CASE("Simple HTML export", "[export][simple]")
+{
+    SECTION("movie") { CHECK(true); }
+}
